AlbumController.java 작성
- submit 작성..
- 로컬?에 있는 image 파일이 삭제된 경우 처리해주기

+++ release/src/main/java/com/project/release/controller/PhotoForm.java
@@ -7,7 +7,6 @@ import lombok.Setter;
 @Setter
 public class PhotoForm {
     private String title;
-    private Integer num;
+    private Long num;
     private String description;
-
 }
